Core: colocate bash unit tests with implementation.
